% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin.R
\name{coef.merlin}
\alias{coef.merlin}
\alias{coef.summary.merlin}
\title{Extract Model Coefficients}
\usage{
\method{coef}{merlin}(object, ...)

\method{coef}{summary.merlin}(object, ...)
}
\arguments{
\item{object}{An object of class \code{merlin} or \code{summary.merlin}.}

\item{...}{Not used.}
}
\description{
\code{coef} extracts model coefficients from a \code{merlin} model fit. \code{coefficients} is an alias for it.
}
