% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrcs.R
\name{coef.mlrcs}
\alias{coef.mlrcs}
\alias{coef.summary.mlrcs}
\title{Extract Model Coefficients}
\usage{
\method{coef}{mlrcs}(object, ...)

\method{coef}{summary.mlrcs}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mlrcs} or \code{summary.mlrcs}.}

\item{...}{Not used.}
}
\description{
\code{coef} extracts model coefficients from a \code{mlrcs} model fit. \code{coefficients} is an alias for it.
}
