% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsurv.R
\name{coef.mlsurv}
\alias{coef.mlsurv}
\alias{coef.summary.mlsurv}
\title{Extract Model Coefficients}
\usage{
\method{coef}{mlsurv}(object, ...)

\method{coef}{summary.mlsurv}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mlsurv} or \code{summary.mlsurv}.}

\item{...}{Not used.}
}
\description{
\code{coef} extracts model coefficients from a \code{mlsurv} model fit. \code{coefficients} is an alias for it.
}
