% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin.R
\name{logLik.merlin}
\alias{logLik.merlin}
\alias{logLik.summary.merlin}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{merlin}(object, ...)

\method{logLik}{summary.merlin}(object, ...)
}
\arguments{
\item{object}{An object of class \code{merlin} or \code{summary.merlin}.}

\item{...}{Not used.}
}
\description{
Extract log-likelihood of a \code{merlin} model.
}
