% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrcs.R
\name{logLik.mlrcs}
\alias{logLik.mlrcs}
\alias{logLik.summary.mlrcs}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{mlrcs}(object, ...)

\method{logLik}{summary.mlrcs}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mlrcs} or \code{summary.mlrcs}.}

\item{...}{Not used.}
}
\description{
Extract log-likelihood of a \code{mlrcs} model.
}
