% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsurv.R
\name{logLik.mlsurv}
\alias{logLik.mlsurv}
\alias{logLik.summary.mlsurv}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{mlsurv}(object, ...)

\method{logLik}{summary.mlsurv}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mlsurv} or \code{summary.mlsurv}.}

\item{...}{Not used.}
}
\description{
Extract log-likelihood of a \code{mlsurv} model.
}
