% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin_utils.R
\name{merlin_util_ap_mod}
\alias{merlin_util_ap_mod}
\title{merlin_util_ap_mod - returns the current estimate of the ith ancillary parameter for the specified model}
\usage{
merlin_util_ap_mod(gml, m, i)
}
\arguments{
\item{gml}{merlin object - should not be edited}

\item{m}{specifies the merlin submodel}

\item{i}{index for which ancillary parameter to extract}
}
\description{
merlin_util_ap_mod - returns the current estimate of the ith ancillary parameter for the specified model
}
\references{
Crowther MJ. Extended multivariate generalised linear and non-linear mixed effects models. \url{https://arxiv.org/abs/1710.02223}

Crowther MJ. merlin - a unified framework for data analysis and methods development in Stata. \url{https://arxiv.org/abs/1806.01615}

Martin EC, Gasparini A, Crowther MJ. merlin - an R package for mixed effects regression of linear and non-linear models.
}
\author{
Michael J. Crowther
}
