% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin_utils.R
\name{merlin_util_depvar}
\alias{merlin_util_depvar}
\title{merlin_util_depvar - returns the response variable(s)}
\usage{
merlin_util_depvar(gml)
}
\arguments{
\item{gml}{merlin object - should not be edited}
}
\description{
Utility function to extract the dependent variable(s) for the current model. If the model
is a survival/time-to-event model, then it will contain the event times in the first column
and the event indicator in the second.
}
\references{
Crowther MJ. Extended multivariate generalised linear and non-linear mixed effects models. \url{https://arxiv.org/abs/1710.02223}

Crowther MJ. merlin - a unified framework for data analysis and methods development in Stata. \url{https://arxiv.org/abs/1806.01615}

Martin EC, Gasparini A, Crowther MJ. merlin - an R package for mixed effects regression of linear and non-linear models.
}
\author{
Emma C. Martin, Alessandro Gasparini and Michael J. Crowther
}
