% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin_utils.R
\name{merlin_util_ev}
\alias{merlin_util_ev}
\title{merlin_util_ev - returns the observation-level expected value of the outcome}
\usage{
merlin_util_ev(gml, t = NULL)
}
\arguments{
\item{gml}{merlin object - should not be edited}

\item{t}{specifies the variable which represents time}
}
\description{
Utility function to extract the expected value of the outcome at the current parameter estimates
for a particular model.
}
\references{
Crowther MJ. Extended multivariate generalised linear and non-linear mixed effects models. \url{https://arxiv.org/abs/1710.02223}

Crowther MJ. merlin - a unified framework for data analysis and methods development in Stata. \url{https://arxiv.org/abs/1806.01615}

Martin EC, Gasparini A, Crowther MJ. merlin - an R package for mixed effects regression of linear and non-linear models.
}
\author{
Michael J. Crowther
}
