% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin.R
\name{print.merlin}
\alias{print.merlin}
\title{Print \code{merlin} Fits}
\usage{
\method{print}{merlin}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{merlin}.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Not used.}
}
\description{
Print the coefficients from a \code{merlin} fit.
}
