% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrcs.R
\name{print.mlrcs}
\alias{print.mlrcs}
\title{Print \code{mlrcs} Fits}
\usage{
\method{print}{mlrcs}(x, digits = max(3L, getOption("digits") - 3L), simplify = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{mlrcs}.}

\item{digits}{The number of significant digits to use when printing.}

\item{simplify}{Should non-interpretable coefficients be hidden (e.g. splines and flexible polynomials terms)? Defaults to TRUE.}

\item{...}{Not used.}
}
\description{
Print the coefficients from a \code{mlrsc} fit.
}
