% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsurv.R
\name{print.mlsurv}
\alias{print.mlsurv}
\title{Print \code{mlsurv} Fits}
\usage{
\method{print}{mlsurv}(x, digits = max(3L, getOption("digits") - 3L), simplify = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{mlsurv}.}

\item{digits}{The number of significant digits to use when printing.}

\item{simplify}{Should non-interpretable coefficients be hidden (e.g. splines and flexible polynomials terms)? Defaults to TRUE.}

\item{...}{Not used.}
}
\description{
Print the coefficients from a \code{mlsurv} fit.
}
