% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin.R
\name{summary.merlin}
\alias{summary.merlin}
\alias{print.summary.merlin}
\title{Summarizing \code{merlin} Fits}
\usage{
\method{summary}{merlin}(object, sig = 0.95, ...)

\method{print}{summary.merlin}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{An object of class \code{merlin}}

\item{sig}{Significancy level for confidence intervals. Defaults to 0.95.}

\item{...}{Not used.}

\item{x}{An object of class \code{summary.merlin}}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
These functions are all methods for class \code{merlin} or \code{summary.merlin} objects.
}
