% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrcs.R
\name{summary.mlrcs}
\alias{summary.mlrcs}
\alias{print.summary.mlrcs}
\title{Summarizing \code{mlrcs} Fits}
\usage{
\method{summary}{mlrcs}(object, sig = 0.95, ...)

\method{print}{summary.mlrcs}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{An object of class \code{mlrcs}}

\item{sig}{Significancy level for confidence intervals. Defaults to 0.95.}

\item{...}{Not used.}

\item{x}{An object of class \code{summary.mlrcs}}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
These functions are all methods for class \code{mlrcs} or \code{summary.mlrcs} objects.
}
