% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsurv.R
\name{summary.mlsurv}
\alias{summary.mlsurv}
\alias{print.summary.mlsurv}
\title{Summarizing \code{mlsurv} Fits}
\usage{
\method{summary}{mlsurv}(object, sig = 0.95, ...)

\method{print}{summary.mlsurv}(x, digits = max(3, getOption("digits") - 3), simplify = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{mlsurv}}

\item{sig}{Significancy level for confidence intervals. Defaults to 0.95.}

\item{...}{Not used.}

\item{x}{An object of class \code{summary.mlsurv}}

\item{digits}{The number of significant digits to use when printing.}

\item{simplify}{Should non-interpretable coefficients be hidden (e.g. splines and flexible polynomials terms)? Defaults to TRUE.}
}
\description{
These functions are all methods for class \code{mlsurv} or \code{summary.mlsurv} objects.
}
