% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merlin.R
\name{vcov.merlin}
\alias{vcov.merlin}
\alias{vcov.summary.merlin}
\title{Calculate Variance-Covariance Matrix for a \code{merlin} Model Object}
\usage{
\method{vcov}{merlin}(object, ...)

\method{vcov}{summary.merlin}(object, ...)
}
\arguments{
\item{object}{An object of class \code{merlin} or \code{summary.merlin}.}

\item{...}{Not used.}
}
\description{
Returns the variance-covariance matrix of all estimated parameters of a fitted \code{merlin} model.
}
