% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrcs.R
\name{vcov.mlrcs}
\alias{vcov.mlrcs}
\alias{vcov.summary.mlrcs}
\title{Calculate Variance-Covariance Matrix for a \code{mlrcs} Model Object}
\usage{
\method{vcov}{mlrcs}(object, ...)

\method{vcov}{summary.mlrcs}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mlrcs} or \code{summary.mlrcs}.}

\item{...}{Not used.}
}
\description{
Returns the variance-covariance matrix of all estimated parameters of a fitted \code{mlrcs} model.
}
