% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsurv.R
\name{vcov.mlsurv}
\alias{vcov.mlsurv}
\alias{vcov.summary.mlsurv}
\title{Calculate Variance-Covariance Matrix for a \code{mlsurv} Model Object}
\usage{
\method{vcov}{mlsurv}(object, ...)

\method{vcov}{summary.mlsurv}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mlsurv} or \code{summary.mlsurv}.}

\item{...}{Not used.}
}
\description{
Returns the variance-covariance matrix of all estimated parameters of a fitted \code{mlsurv} model.
}
