% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERLIN-package.R
\docType{package}
\name{MERLIN-package}
\alias{MERLIN}
\alias{MERLIN-package}
\title{MERLIN: MERLIN for causal heterogeneity using summary statistics}
\description{
MERLIN (Causal Heterogeneity using Summary Statistics) is a unified Bayesian framework that jointly estimates average and heterogeneity effects using summary data from genome-wide association studies (GWAS) and Genome-Wide Interaction Studies (GWIS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shilab-ecnu/MERLIN}
  \item \url{https://shilab-ecnu.github.io/MERLIN/}
  \item Report bugs at \url{https://github.com/shilab-ecnu/MERLIN/issues/}
}

}
\author{
\strong{Maintainer}: Jie He \email{10225000415@stu.ecnu.edu.cn}

Authors:
\itemize{
  \item Yadong Yang \email{52294404009@stu.ecnu.edu.cn}
  \item Xingjie Shi \email{xjshi@fem.ecnu.edu.cn} (\href{https://orcid.org/0000-0002-9866-8599}{ORCID})
  \item Minxi Bai \email{mxbai0417@sjtu.edu.cn}
  \item Jie He \email{10225000415@stu.ecnu.edu.cn}
}

}
\keyword{internal}
